Make scheduling configuration compatible with musl
Musl does not implement sched_setscheduler() because it operates on a
thread, while POSIX requires it to operate on the process [1]. We only
need to change priority of the current thread, so using
pthread_setschedparam() instead works with no disadvantage.

Additional musl's sched_param has additional fields for the sporadic
server scheduler, which must be filled even if we don't use that
scheduler.

[1] https://www.openwall.com/lists/musl/2016/03/01/4